requirements.txt
speedtest == 0.0.1
psycopg2-binary == 2.8.6